+++ packable/packable-derive-test/tests/pass/verify_with_field.rs
@@ -21,12 +21,8 @@ impl From<Infallible> for PickyError {
     }
 }
 
-fn verify_value<const VERIFY: bool>(&value: &u8) -> Result<(), PickyError> {
-    if !VERIFY || value == 42 {
-        Ok(())
-    } else {
-        Err(PickyError(value))
-    }
+fn verify_value(&value: &u8) -> Result<(), PickyError> {
+    if value == 42 { Ok(()) } else { Err(PickyError(value)) }
 }
 
 #[derive(Packable)]