+++ packable/packable-derive/src/fragments.rs
@@ -32,21 +32,39 @@ impl Fragments {
             fields_type,
         } = info;
 
-        let fields_verification = fields_verify_with.into_iter().zip(fields_ident.iter()).map(|(verify_with, field_ident)| match verify_with {
-            Some(verify_with) => if unpack_visitor_info.explicit {
-                quote!(#verify_with::<VERIFY>(&#field_ident, visitor).map_err(#crate_name::error::UnpackError::from_packable)?;)
-            } else {
-                quote!(#verify_with::<VERIFY>(&#field_ident).map_err(#crate_name::error::UnpackError::from_packable)?;)
-            }
-            None => quote!(),
-        });
+        let fields_verification = fields_verify_with.into_iter().zip(fields_ident.iter()).map(
+            |(verify_with, field_ident)| match verify_with {
+                Some(verify_with) => if unpack_visitor_info.explicit {
+                    quote! {
+                        if let Some(visitor) = visitor {
+                            #verify_with(&#field_ident, visitor).map_err(#crate_name::error::UnpackError::from_packable)?;
+                        }
+                    }
+                } else {
+                    quote! {
+                        if visitor.is_some() {
+                            #verify_with(&#field_ident).map_err(#crate_name::error::UnpackError::from_packable)?;
+                        }
+                    }
+                },
+                None => quote!(),
+            },
+        );
 
         let verify_with = match verify_with {
             Some(verify_with) => {
                 if unpack_visitor_info.explicit {
-                    quote!(#verify_with::<VERIFY>(&unpacked, visitor).map_err(#crate_name::error::UnpackError::from_packable)?;)
+                    quote! {
+                        if let Some(visitor) = visitor {
+                            #verify_with(&unpacked, visitor).map_err(#crate_name::error::UnpackError::from_packable)?;
+                        }
+                    }
                 } else {
-                    quote!(#verify_with::<VERIFY>(&unpacked).map_err(#crate_name::error::UnpackError::from_packable)?;)
+                    quote! {
+                        if visitor.is_some() {
+                            #verify_with(&unpacked).map_err(#crate_name::error::UnpackError::from_packable)?;
+                        }
+                    }
                 }
             }
             None => quote!(),
@@ -60,7 +78,7 @@ impl Fragments {
             },
             unpack: quote! {
                 #(
-                    let #fields_ident = <#fields_type as #crate_name::Packable>::unpack::<_, VERIFY>(unpacker, Borrow::<<#fields_type as #crate_name::Packable>::UnpackVisitor>::borrow(visitor)).map_packable_err(#fields_unpack_error_with).coerce()?;
+                    let #fields_ident = <#fields_type as #crate_name::Packable>::unpack(unpacker, visitor.map(Borrow::<<#fields_type as #crate_name::Packable>::UnpackVisitor>::borrow)).map_packable_err(#fields_unpack_error_with).coerce()?;
                     #fields_verification
                 )*