+++ packable/packable/src/packable/box.rs
@@ -22,11 +22,11 @@ impl<T: Packable> Packable for Box<T> {
     }
 
     #[inline]
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
         unpacker: &mut U,
-        visitor: &Self::UnpackVisitor,
+        visitor: Option<&Self::UnpackVisitor>,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
-        Ok(Box::new(T::unpack::<_, VERIFY>(unpacker, visitor)?))
+        Ok(Box::new(T::unpack(unpacker, visitor)?))
     }
 }
 
@@ -54,13 +54,13 @@ impl<T: Packable> Packable for Box<[T]> {
     }
 
     #[inline]
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
         unpacker: &mut U,
-        visitor: &Self::UnpackVisitor,
+        visitor: Option<&Self::UnpackVisitor>,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
         use crate::error::UnpackErrorExt;
 
-        let len = u64::unpack::<_, VERIFY>(unpacker, &())
+        let len = u64::unpack_inner(unpacker, visitor)
             .coerce()?
             .try_into()
             .map_err(|err| UnpackError::Packable(Self::UnpackError::Prefix(err)))?;
@@ -74,7 +74,7 @@ impl<T: Packable> Packable for Box<[T]> {
             let mut vec = Vec::with_capacity(len);
 
             for _ in 0..len {
-                let item = T::unpack::<_, VERIFY>(unpacker, visitor).map_packable_err(Self::UnpackError::Item)?;
+                let item = T::unpack(unpacker, visitor).map_packable_err(Self::UnpackError::Item)?;
                 vec.push(item);
             }