+++ packable/packable/src/packable/map.rs
@@ -137,13 +137,13 @@ where
     }
 
     #[inline]
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
         unpacker: &mut U,
-        visitor: &Self::UnpackVisitor,
+        visitor: Option<&Self::UnpackVisitor>,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
         use crate::error::UnpackErrorExt;
 
-        let len = u64::unpack::<_, VERIFY>(unpacker, &())
+        let len = u64::unpack_inner(unpacker, visitor)
             .coerce()?
             .try_into()
             .map_err(|err| UnpackError::Packable(UnpackMapError::Prefix(err)))?;
@@ -151,7 +151,7 @@ where
         let mut map = HashMap::<K, V>::with_capacity(len);
 
         for _ in 0..len {
-            let key = K::unpack::<_, VERIFY>(unpacker, visitor.borrow())
+            let key = K::unpack_inner(unpacker, visitor)
                 .map_packable_err(UnpackMapError::Key)
                 .map_packable_err(Self::UnpackError::from)?;
 
@@ -159,7 +159,7 @@ where
                 return Err(UnpackError::Packable(UnpackMapError::DuplicateKey(key)));
             }
 
-            let value = V::unpack::<_, VERIFY>(unpacker, visitor)
+            let value = V::unpack(unpacker, visitor)
                 .map_packable_err(UnpackMapError::Value)
                 .map_packable_err(Self::UnpackError::from)?;
 
@@ -192,13 +192,13 @@ where
     }
 
     #[inline]
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
         unpacker: &mut U,
-        visitor: &Self::UnpackVisitor,
+        visitor: Option<&Self::UnpackVisitor>,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
         use crate::error::UnpackErrorExt;
 
-        let len = u64::unpack::<_, VERIFY>(unpacker, &())
+        let len = u64::unpack_inner(unpacker, visitor)
             .coerce()?
             .try_into()
             .map_err(|err| UnpackError::Packable(UnpackMapError::Prefix(err).into()))?;
@@ -206,7 +206,7 @@ where
         let mut map = BTreeMap::<K, V>::new();
 
         for _ in 0..len {
-            let key = K::unpack::<_, VERIFY>(unpacker, visitor.borrow())
+            let key = K::unpack_inner(unpacker, visitor)
                 .map_packable_err(UnpackMapError::Key)
                 .map_packable_err(Self::UnpackError::from)?;
 
@@ -224,7 +224,7 @@ where
                 }
             }
 
-            let value = V::unpack::<_, VERIFY>(unpacker, visitor)
+            let value = V::unpack(unpacker, visitor)
                 .map_packable_err(UnpackMapError::Value)
                 .map_packable_err(Self::UnpackError::from)?;