+++ packable/packable/src/packable/mod.rs
@@ -80,15 +80,13 @@ use crate::{
 ///         }
 ///     }
 ///
-///     fn unpack<U: Unpacker, const VERIFY: bool>(
+///     fn unpack<U: Unpacker>(
 ///         unpacker: &mut U,
-///         visitor: &Self::UnpackVisitor,
+///         visitor: Option<&Self::UnpackVisitor>,
 ///     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
-///         match u8::unpack::<_, VERIFY>(unpacker, visitor).coerce()? {
+///         match u8::unpack_inner(unpacker, visitor).coerce()? {
 ///             0u8 => Ok(Self::Nothing),
-///             1u8 => Ok(Self::Just(
-///                 i32::unpack::<_, VERIFY>(unpacker, visitor).coerce()?,
-///             )),
+///             1u8 => Ok(Self::Just(i32::unpack_inner(unpacker, visitor).coerce()?)),
 ///             tag => Err(UnpackError::Packable(UnknownTagError(tag))),
 ///         }
 ///     }
@@ -174,11 +172,11 @@ use crate::{
 /// `#[packable(verify_with = ...)]` attribute which must receive a valid Rust path refering to a
 /// function with the signature
 /// ```ignore
-/// fn<const VERIFY: bool>(field: &F) -> Result<(), P::UnpackError>
+/// fn<P: Packable, F>(field: &F, visitor: &P::UnpackVisitor) -> Result<(), P::UnpackError>
 /// ```
-/// where `F` is the type of the field being verified, `P` is the type of the `struct` or `enum`
-/// and `VERIFY` is the same constant parameter used inside `Packable::unpack`. This verification
-/// function will be run immediately after unpacking the field.
+/// where `F` is the type of the field being verified, `P` is the type of the `struct` or `enum`.
+/// This verification function will be run immediately after unpacking the field. The `visitor`
+/// param can be excluded if it is not needed for verification.
 pub trait Packable: Sized + 'static {
     /// The error type that can be returned if some semantic error occurs while unpacking.
     ///
@@ -186,17 +184,38 @@ pub trait Packable: Sized + 'static {
     /// [`UnknownTagError`](crate::error::UnknownTagError) when implementing this trait for an enum.
     type UnpackError: Debug + From<Infallible>;
     /// FIXME: docs
-    type UnpackVisitor: Default + Borrow<()>;
+    type UnpackVisitor: Borrow<()>;
 
     /// Packs this value into the given [`Packer`].
     fn pack<P: Packer>(&self, packer: &mut P) -> Result<(), P::Error>;
 
-    /// Unpacks this value from the given [`Unpacker`]. The `VERIFY` generic parameter can be used to skip additional
+    /// Unpacks this value from the given [`Unpacker`]. The optional visitor can be used to make additional
     /// syntactic checks.
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
+        unpacker: &mut U,
+        visitor: Option<&Self::UnpackVisitor>,
+    ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>>;
+
+    /// Unpacks this value from the given [`Unpacker`] with a visitor for validation.
+    fn unpack_verified<U: Unpacker>(
         unpacker: &mut U,
         visitor: &Self::UnpackVisitor,
-    ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>>;
+    ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
+        Self::unpack(unpacker, Some(visitor))
+    }
+
+    /// Unpacks this value from the given [`Unpacker`] without validation.
+    fn unpack_unverified<U: Unpacker>(unpacker: &mut U) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
+        Self::unpack(unpacker, None)
+    }
+
+    /// Unpacks an inner value whose visitor type can be borrowed from the outer visitor.
+    fn unpack_inner<V: Borrow<Self::UnpackVisitor>, U: Unpacker>(
+        unpacker: &mut U,
+        visitor: Option<&V>,
+    ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
+        Self::unpack(unpacker, visitor.map(Borrow::borrow))
+    }
 }
 
 /// Extension trait for types that implement [`Packable`].
@@ -209,13 +228,13 @@ pub trait PackableExt: Packable {
     fn pack_to_vec(&self) -> Vec<u8>;
 
     /// Unpacks this value from a sequence of bytes doing syntactical checks.
-    fn unpack_verified<T: AsRef<[u8]>>(
+    fn unpack_bytes_verified<T: AsRef<[u8]>>(
         bytes: T,
         visitor: &Self::UnpackVisitor,
     ) -> Result<Self, UnpackError<<Self as Packable>::UnpackError, UnexpectedEOF>>;
 
     /// Unpacks this value from a sequence of bytes without doing syntactical checks.
-    fn unpack_unverified<T: AsRef<[u8]>>(
+    fn unpack_bytes_unverified<T: AsRef<[u8]>>(
         bytes: T,
     ) -> Result<Self, UnpackError<<Self as Packable>::UnpackError, UnexpectedEOF>>;
 }
@@ -243,21 +262,18 @@ impl<P: Packable> PackableExt for P {
 
     /// Unpacks this value from a type that implements [`AsRef<[u8]>`].
     #[inline]
-    fn unpack_verified<T: AsRef<[u8]>>(
+    fn unpack_bytes_verified<T: AsRef<[u8]>>(
         bytes: T,
         visitor: &P::UnpackVisitor,
     ) -> Result<Self, UnpackError<<Self as Packable>::UnpackError, UnexpectedEOF>> {
-        Self::unpack::<_, true>(&mut SliceUnpacker::new(bytes.as_ref()), visitor)
+        Self::unpack_verified(&mut SliceUnpacker::new(bytes.as_ref()), visitor)
     }
 
     /// Unpacks this value from a type that implements [`AsRef<[u8]>`] skipping some syntatical checks.
     #[inline]
-    fn unpack_unverified<T: AsRef<[u8]>>(
+    fn unpack_bytes_unverified<T: AsRef<[u8]>>(
         bytes: T,
     ) -> Result<Self, UnpackError<<Self as Packable>::UnpackError, UnexpectedEOF>> {
-        Self::unpack::<_, false>(
-            &mut SliceUnpacker::new(bytes.as_ref()),
-            &<P as Packable>::UnpackVisitor::default(),
-        )
+        Self::unpack_unverified(&mut SliceUnpacker::new(bytes.as_ref()))
     }
 }