+++ packable/packable/src/packable/option.rs
@@ -54,14 +54,14 @@ impl<T: Packable> Packable for Option<T> {
         }
     }
 
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
         unpacker: &mut U,
-        visitor: &Self::UnpackVisitor,
+        visitor: Option<&Self::UnpackVisitor>,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
-        match u8::unpack::<_, VERIFY>(unpacker, &()).coerce()? {
+        match u8::unpack_inner(unpacker, visitor).coerce()? {
             0 => Ok(None),
             1 => Ok(Some(
-                T::unpack::<_, VERIFY>(unpacker, visitor).map_packable_err(UnpackOptionError::Inner)?,
+                T::unpack(unpacker, visitor).map_packable_err(UnpackOptionError::Inner)?,
             )),
             n => Err(UnpackError::Packable(Self::UnpackError::UnknownTag(n))),
         }