+++ packable/packable/src/packable/prefix/string.rs
@@ -101,12 +101,12 @@ where
     }
 
     #[inline]
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
         unpacker: &mut U,
-        visitor: &Self::UnpackVisitor,
+        visitor: Option<&Self::UnpackVisitor>,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
         // The length of any dynamically-sized sequence must be prefixed.
-        let len = B::unpack::<_, VERIFY>(unpacker, visitor)
+        let len = B::unpack(unpacker, visitor)
             .map_packable_err(UnpackPrefixError::Prefix)?
             .into();