+++ packable/packable/src/packable/string.rs
@@ -32,11 +32,11 @@ impl Packable for String {
     }
 
     #[inline]
-    fn unpack<U: Unpacker, const VERIFY: bool>(
+    fn unpack<U: Unpacker>(
         unpacker: &mut U,
-        visitor: &Self::UnpackVisitor,
+        visitor: Option<&Self::UnpackVisitor>,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
-        let bytes = Vec::<u8>::unpack::<_, VERIFY>(unpacker, visitor)
+        let bytes = Vec::<u8>::unpack(unpacker, visitor)
             .map_packable_err(|err| UnpackPrefixError::Prefix(err.into_prefix_err()))?;
 
         String::from_utf8(bytes).map_err(|e| UnpackError::Packable(Self::UnpackError::Item(e)))