+++ packable/packable/tests/btreemap.rs
@@ -34,7 +34,7 @@ fn invalid_duplicate() {
             .chain(bytes.into_iter().flat_map(|(k, v)| [k, v])),
     );
 
-    let prefixed = BTreeMap::<u8, u8>::unpack_verified(bytes, &());
+    let prefixed = BTreeMap::<u8, u8>::unpack_bytes_verified(bytes, &());
 
     assert!(matches!(
         prefixed,
@@ -55,7 +55,7 @@ fn invalid_unordered() {
             .chain(bytes.into_iter().flat_map(|(k, v)| [k, v])),
     );
 
-    let prefixed = BTreeMap::<u8, u8>::unpack_verified(bytes, &());
+    let prefixed = BTreeMap::<u8, u8>::unpack_bytes_verified(bytes, &());
 
     assert!(matches!(
         prefixed,