+++ packable/packable/tests/btreemap_prefix.rs
@@ -63,7 +63,7 @@ macro_rules! impl_packable_test_for_btreemap_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeMapPrefix::<u8, u8, $ty>::unpack_verified(bytes, &());
+            let prefixed = BTreeMapPrefix::<u8, u8, $ty>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -85,7 +85,7 @@ macro_rules! impl_packable_test_for_btreemap_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeMapPrefix::<u8, u8, $ty>::unpack_verified(bytes, &());
+            let prefixed = BTreeMapPrefix::<u8, u8, $ty>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -128,7 +128,7 @@ macro_rules! impl_packable_test_for_bounded_btreemap_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(core::iter::repeat(0).take(2 * (LEN + 1))));
 
-            let prefixed = BTreeMapPrefix::<u8, u8, $bounded<$min, $max>>::unpack_verified(bytes, &());
+            let prefixed = BTreeMapPrefix::<u8, u8, $bounded<$min, $max>>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -149,7 +149,7 @@ macro_rules! impl_packable_test_for_bounded_btreemap_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeMapPrefix::<u8, u8, $bounded<$min, $max>>::unpack_verified(bytes, &());
+            let prefixed = BTreeMapPrefix::<u8, u8, $bounded<$min, $max>>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -171,7 +171,7 @@ macro_rules! impl_packable_test_for_bounded_btreemap_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeMapPrefix::<u8, u8, $bounded<$min, $max>>::unpack_verified(bytes, &());
+            let prefixed = BTreeMapPrefix::<u8, u8, $bounded<$min, $max>>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,