+++ packable/packable/tests/btreeset.rs
@@ -26,7 +26,7 @@ fn invalid_duplicate() {
     let bytes = [1, 2, 3, 3, 4];
     let bytes = Vec::from_iter(bytes.len().to_le_bytes().into_iter().chain(bytes));
 
-    let prefixed = BTreeSet::<u8>::unpack_verified(bytes, &());
+    let prefixed = BTreeSet::<u8>::unpack_bytes_verified(bytes, &());
 
     println!("{prefixed:?}");
 
@@ -43,7 +43,7 @@ fn invalid_unordered() {
     let bytes = [1, 2, 4, 3];
     let bytes = Vec::from_iter(bytes.len().to_le_bytes().into_iter().chain(bytes));
 
-    let prefixed = BTreeSet::<u8>::unpack_verified(bytes, &());
+    let prefixed = BTreeSet::<u8>::unpack_bytes_verified(bytes, &());
 
     assert!(matches!(
         prefixed,