+++ packable/packable/tests/btreeset_prefix.rs
@@ -63,7 +63,7 @@ macro_rules! impl_packable_test_for_btreeset_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeSetPrefix::<u8, $ty>::unpack_verified(bytes, &());
+            let prefixed = BTreeSetPrefix::<u8, $ty>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -83,7 +83,7 @@ macro_rules! impl_packable_test_for_btreeset_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeSetPrefix::<u8, $ty>::unpack_verified(bytes, &());
+            let prefixed = BTreeSetPrefix::<u8, $ty>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -126,7 +126,7 @@ macro_rules! impl_packable_test_for_bounded_btreeset_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(core::iter::repeat(0).take(LEN + 1)));
 
-            let prefixed = BTreeSetPrefix::<u8, $bounded<$min, $max>>::unpack_verified(bytes, &());
+            let prefixed = BTreeSetPrefix::<u8, $bounded<$min, $max>>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -147,7 +147,7 @@ macro_rules! impl_packable_test_for_bounded_btreeset_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeSetPrefix::<u8, $bounded<$min, $max>>::unpack_verified(bytes, &());
+            let prefixed = BTreeSetPrefix::<u8, $bounded<$min, $max>>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,
@@ -167,7 +167,7 @@ macro_rules! impl_packable_test_for_bounded_btreeset_prefix {
 
             let bytes = Vec::from_iter(LEN_AS_TY.to_le_bytes().into_iter().chain(bytes));
 
-            let prefixed = BTreeSetPrefix::<u8, $bounded<$min, $max>>::unpack_verified(bytes, &());
+            let prefixed = BTreeSetPrefix::<u8, $bounded<$min, $max>>::unpack_bytes_verified(bytes, &());
 
             assert!(matches!(
                 prefixed,